requirements.txt
# DA6401 Assignment 2 — permitted libraries only
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=9.5.0
scikit-learn>=1.2.0
wandb>=0.15.0
albumentations>=1.3.0
gdown>=4.7.1
